some table view tweaks

+++ common/src/main/kotlin/com/metacontent/cobblenav/client/gui/widget/fishing/BucketViewWidget.kt
@@ -19,8 +19,8 @@ class BucketViewWidget(
     width: Int,
     columns: Int,
     columnWidth: Int = width / columns,
-    verticalPadding: Int = 0,
-    horizontalPadding: Int = (width - columns * columnWidth) / (columns - 1),
+    verticalPadding: Float = 0f,
+    horizontalPadding: Float = (width - columns * columnWidth) / (columns - 1f),
     val minHeight: Int,
     val bucket: WeightedBucket
 ) : TableView<ScrollableItemWidget<SpawnDataWidget>>(
@@ -36,8 +36,8 @@ class BucketViewWidget(
         const val SEPARATOR_HEIGHT = 4
         const val BUCKET_PADDING = 4
         const val BUCKET_WIDTH = 60
-        val UP_COLOR = RGB(0,64,128)
-        val DOWN_COLOR = RGB(0,13,25)
+        val UP_COLOR = RGB(0, 64, 128)
+        val DOWN_COLOR = RGB(0, 13, 25)
     }
 
     val color = interpolate(DOWN_COLOR, UP_COLOR, sqrt(bucket.chance)).toColor(200)
